requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.4
pydantic==1.10.7
python-dotenv==1.0.0
gunicorn==20.1.0
httpx==0.24.0
pytest==7.3.1
python-jose==3.3.0
passlib==1.7.4
uuid==1.30 
sqlalchemy